fix: Skip maintainer updates that conflict
We're using the github_id as a primary key for the NixMaintainer model
which prevents us from updating this field for a specific username and
we can't have duplicate usernames either since we have a unique
contstraint on them.

This "patches" #658 so ingestion works, until we figure out how to
handle #657.